# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Campaign:
    properties:
      category_id:
        type: integer
      company:
        $ref: '#/definitions/entity.Company'
        type: object
      company_id:
        type: integer
      currency:
        type: string
      detail:
        type: string
      end_date:
        type: string
      end_price:
        type: number
      id:
        type: integer
      location:
        $ref: '#/definitions/entity.Location'
        type: object
      location_id:
        type: integer
      payment_days:
        $ref: '#/definitions/entity.PaymentDays'
        type: object
      payment_days_id:
        type: integer
      payment_terms:
        $ref: '#/definitions/entity.PaymentTerms'
        type: object
      payment_terms_id:
        type: integer
      staff_amount:
        type: integer
      start_date:
        type: string
      start_price:
        type: number
      status:
        type: string
      sub_category_id:
        type: integer
      title:
        type: string
    required:
    - currency
    - detail
    - end_price
    - staff_amount
    - start_price
    - title
    type: object
...
  entity.Category:
    properties:
      is_searchable:
        type: boolean
      is_usable:
        type: boolean
      name:
        type: string
      slug:
        type: string
      sub_categories:
        items:
          $ref: '#/definitions/entity.SubCategory'
        type: array
    type: object
  entity.Company:
    properties:
      background_image:
        $ref: '#/definitions/entity.Image'
        type: object
      category:
        $ref: '#/definitions/entity.Category'
        type: object
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      image:
        $ref: '#/definitions/entity.Image'
        type: object
      is_updated:
        type: boolean
      is_verified:
        type: boolean
      name:
        type: string
      phone:
        type: string
      sub_category:
        $ref: '#/definitions/entity.SubCategory'
        type: object
      sub_category_id:
        type: integer
    required:
    - name
    type: object
...
  entity.Expertise:
    properties:
      is_searchable:
        type: boolean
      is_usable:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
...
  entity.Image:
    properties:
      id:
        type: integer
      image_url:
        type: string
    required:
    - image_url
    type: object
...
  entity.Location:
    properties:
      address:
        type: string
      formatted_address:
        type: string
      icon:
        type: string
      id:
        type: integer
      international_phone_number:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      phone_number:
        type: string
      place_id:
        type: string
    required:
    - formatted_address
    - lat
    - lng
    type: object
  entity.Occupation:
    properties:
      is_searchable:
        type: boolean
      is_usable:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
...
  entity.PaymentDays:
    properties:
      days:
        type: integer
      id:
        type: integer
    required:
    - days
    type: object
  entity.PaymentTerms:
    properties:
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    required:
    - name
    type: object
  entity.Portfolio:
    properties:
      id:
        type: integer
      image_url:
        type: string
      service_id:
        type: integer
      title:
        type: string
    type: object
...
  entity.Service:
    properties:
      backgrounds:
        items:
          $ref: '#/definitions/entity.Image'
        type: array
      category:
        $ref: '#/definitions/entity.Category'
        type: object
      category_id:
        type: integer
      cpv:
        type: number
      description:
        type: string
      id:
        type: integer
      language:
        type: string
      portfolios:
        items:
          $ref: '#/definitions/entity.Portfolio'
        type: array
      price:
        type: number
      status:
        type: string
      sub_category:
        $ref: '#/definitions/entity.SubCategory'
        type: object
      sub_category_id:
        type: integer
      title:
        type: string
      topics:
        items:
          $ref: '#/definitions/entity.Expertise'
        type: array
      user:
        $ref: '#/definitions/entity.User'
        type: object
      user_id:
        type: integer
    required:
    - description
    - language
    - title
    type: object
...
  entity.SubCategory:
    properties:
      category_id:
        type: integer
      is_searchable:
        type: boolean
      is_usable:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
  entity.Talent:
    properties:
      background:
        $ref: '#/definitions/entity.Image'
        type: object
      background_image_id:
        type: integer
      birth_date:
        type: string
      business_type_id:
        type: integer
      description:
        type: string
      engagement:
        type: number
      expertises:
        items:
          $ref: '#/definitions/entity.Expertise'
        type: array
      facebook_followers_count:
        type: integer
      facebook_link:
        type: string
      gender:
        type: string
      height:
        type: integer
      id:
        type: integer
      image:
        $ref: '#/definitions/entity.Image'
        type: object
      instagram_followers_count:
        type: integer
      instagram_link:
        type: string
      is_verified:
        type: boolean
      location:
        $ref: '#/definitions/entity.Location'
        type: object
      location_id:
        type: integer
      occupations:
        items:
          $ref: '#/definitions/entity.Occupation'
        type: array
      phone:
        type: string
      twitter_followers_count:
        type: integer
      twitter_link:
        type: string
      youtube_followers_count:
        type: integer
      youtube_link:
        type: string
    required:
    - birth_date
    - gender
    - height
    - instagram_link
    type: object
...
  entity.User:
    properties:
      active:
        type: boolean
      background_image_url:
        type: string
      currency:
        type: string
      email:
        type: string
      image_url:
        type: string
      language:
        type: string
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/entity.Service'
        type: array
      talent:
        $ref: '#/definitions/entity.Talent'
        type: object
      type:
        type: string
    required:
    - currency
    - email
    - language
    - name
    - type
    type: object
  entity.UserForm:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      type:
        type: string
    required:
    - email
    - name
    - password
    - type
    type: object
...
  translate.ResultErrors:
    properties:
      message:
        example: Dynamic message
        type: string
      status:
        example: Bad Request
        type: string
      status_code:
        example: 400
        type: integer
    type: object
  translate.ResultSuccess:
    properties:
      data:
        type: object
      status:
        example: OK
        type: string
      status_code:
        example: 200
        type: integer
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is event staffing application management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Spgku Application
  version: "1.0"
paths:
  /account:
    get:
      consumes:
      - application/json
      description: Find user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: FindUserByLoggedIn api used to find user by token
      tags:
      - users
  /account/{id}:
    get:
      consumes:
      - application/json
      description: Find user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: FindUserByID api used to find user by primary key
      tags:
      - users
  /auth/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Sign in by using email and password
      parameters:
      - description: your account email
        in: query
        name: email
        required: true
        type: string
      - description: your account password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.Campaign'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/translate.ResultErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      summary: Sign In
      tags:
      - users
...
  /user:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: New User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      summary: AddUser api used to signup
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Edit user
      parameters:
      - description: Edit User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: EditUser api used to edit profile
      tags:
      - users
  /users:
    get:
      consumes:
      - application/json
      description: find users
      parameters:
      - in: query
        name: campaign_id
        type: integer
      - in: query
        name: category_id
        type: integer
      - in: query
        name: expertise_name
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: q
        type: string
      - in: query
        name: sub_category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.User'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: GetUsers is api to find users by service params
      tags:
      - talents
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"